Add support for building kernels for all devices
Currently, prebuilt kernel images are used for most devices. There was
code in place for marlin to rebuild the kernel. This extends the same
flow to all devices. The feature is hidden behind the ENABLE_KERNEL_BUILD
flag. After sufficient testing, it can be hooked to a config option.
Eventually, this would also make it easier to support kernel patches.